+++ NEWS.md
@@ -1,3 +1,41 @@
+Version 0.2.9:  March 21, 2018
+-------------------------------------------------------------------------------
+
+General:
+
++ Added explicit type casting for known columns to `readChangeoDb`.
++ Added the `padSeqEnds` function which pads sequences with Ns to make
+  then equal in length.
++ Added verification of unique sequence IDs to `collapseDuplicates`.
+
+Diversity:
+
++ Added the `uniform` argument to `rarefyDiversity` allowing users to toggle
+  uniform vs non-uniform sampling.
++ Renamed `plotAbundance` to `plotAbundanceCurve`.
++ Changed `estimateAbundance` return object from a data.frame to a new 
+  `AbundanceCurve` custom class.
++ Set default `plot` call for `AbundanceCurve` to `plotAbundanceCurve`.
++ Added the `annotate` argument from `plotDiversityCurve` to 
+  `plotAbundanceCurve`.
++ Added the `score` argument to  `plotDiversityCurve` to toggle between 
+  plotting diversity or evenness.
++ Added the function `plotDiversityTest` to generate a simple plot of
+  `DiversityTest` object summaries.
+
+Gene Usage:
+
++ Added the `omit_nl` argument to `getAllele`, `getGene` and `getFamily` to
+  allow optional filtering of non-localized (NL) genes.
+
+Lineage:
+
++ Fixed a bug in `makeChangeoClone` preventing it from interpreting the `id`
+  argument correctly.
++ Added the `pad_end` argument to `makeChangeoClone` to allow automatic 
+  padding of ends to make sequences the same length.
+
+
 Version 0.2.8:  September 21, 2017
 -------------------------------------------------------------------------------
 
@@ -8,7 +46,6 @@ General:
   sequences but not remove them when set to `TRUE`.
 + Fixed a bug where `collapseDuplicates` was returning one sequence if all 
   sequences were considered ambiguous.
-
 
 Lineage: